+++ baiducloud/extension_instance.go
@@ -1,0 +1,9 @@
+package baiducloud
+
+const (
+	// start the instance
+	INSTANCE_ACTION_START = "start"
+
+	// stop the instance
+	INSTANCE_ACTION_STOP = "stop"
+)