+++ examples/baiducloud-eip/main.tf
@@ -3,10 +3,19 @@ provider "baiducloud" {}
 resource "baiducloud_eip" "my-eip" {
   name              = var.name
   bandwidth_in_mbps = 100
+
+  # support Prepaid/Postpaid
   payment_timing    = "Postpaid"
 
   # support ByTraffic/ByBandwidth
   billing_method = "ByTraffic"
+
+  # only useful when payment_timing is Prepaid
+  # auto_renew_time_unit support month/year
+  # auto_renew_time_unit = "month"
+  # if auto_renew_time_unit is month, auto_renew_time support 1-9
+  # if auto_renew_time_unit is year, auto_renew_time support 1-3
+  # auto_renew_time = 1
 
   tags = {
     "testKey" = "testValue"