+++ examples/baiducloud-subnet/variables.tf
@@ -17,3 +17,4 @@ variable "subnet_cidr" {
 variable "description" {
   default = "this is created by terraform"
 }
+